Fix Ubuntu 7.10 warnings
[ ] May break build
[ ] Breaks existing applications
[ ] Bugfix
[ ] New Feature
[x] Cleanup
[ ] Optimization
[ ] Documentation

+++ lib/client/frame.cpp
@@ -80,7 +80,7 @@ void Frame::syncReadback()
     _frameData->syncReadback();
 }
 
-float Frame::transmit( net::NodePtr toNode )
+int64_t Frame::transmit( net::NodePtr toNode )
 {
     EQASSERT( _frameData );
     return _frameData->transmit( toNode );
